# docs/swagger.yaml
definitions:
  controller.UserLoginResponse:
    properties:
      status_code:
        type: integer
      status_msg:
        type: string
      token:
        type: string
      user_id:
        type: integer
    type: object
  controller.VideoListResponse:
    properties:
      status_code:
        type: integer
      status_msg:
        type: string
      video_list:
        items:
          $ref: '#/definitions/model.Video'
        type: array
    type: object
  favorite.FavoriteListResponse:
    properties:
      status_code:
        type: integer
      status_msg:
        type: string
      video_list:
        items:
          $ref: '#/definitions/model.Video'
        type: array
    type: object
  model.Response:
    properties:
      status_code:
        type: integer
      status_msg:
        type: string
    type: object
  model.UserInfo:
    properties:
      avatar:
        description: 用户头像
        type: string
      background_image:
        description: 用户背景
        type: string
      favorite_count:
        type: integer
      follow_count:
        type: integer
      follower_count:
        type: integer
      id:
        type: integer
      is_follow:
        type: boolean
      name:
        type: string
      total_favorited:
        type: string
      work_count:
        description: 用户作品数量
        type: integer
    type: object
  model.Video:
    properties:
      author:
        allOf:
        - $ref: '#/definitions/model.UserInfo'
        description: 视频作者的相关信息
      comment_count:
        type: integer
      cover_url:
        type: string
      favorite_count:
        type: integer
      id:
        type: integer
      is_favorite:
        type: boolean
      play_url:
        type: string
      title:
        type: string
    type: object
info:
  contact: {}
paths:
  /douyin/favorite/list/:
    get:
      consumes:
      - application/json
      description: 查看用户所有点赞的视频
      parameters:
      - description: 用户鉴权token
        in: query
        name: token
        required: true
        type: string
      - description: 用户的id
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/favorite.FavoriteListResponse'
      summary: 查看用户所有点赞的视频
      tags:
      - 视频
  /douyin/publish/action/:
    post:
      consumes:
      - multipart/form-data
      description: 投稿首先得鉴权，其次获取用户的上传视频，自动获取封面，上传到云存储，并添加到数据库中
      parameters:
      - description: 视频数据
        in: formData
        name: data
        required: true
        type: file
      - description: 用户鉴权token
        in: formData
        name: token
        required: true
        type: string
      - description: 视频标题
        in: formData
        name: title
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
      summary: 用户投稿
      tags:
      - 视频
  /douyin/publish/list/:
    get:
      consumes:
      - application/json
      description: 首先得鉴权，判断token和user_id是否同一个，是否伪造token，其次返回该用户的数据
      parameters:
      - description: 用户鉴权token
        in: query
        name: token
        required: true
        type: string
      - description: 用户的id
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.VideoListResponse'
      summary: 查看用户所有投稿的视频
      tags:
      - 视频
  /douyin/user/login/:
    post:
      consumes:
      - application/json
      description: 用户登录功能，判断密码是否正确
      parameters:
      - description: 账号
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.UserLoginResponse'
      summary: 用户登录
      tags:
      - 用户
  /douyin/user/register/:
    post:
      consumes:
      - application/json
      description: 用户注册，需要判断用户名是否已经被注册了，以及用户密码是否规范
      parameters:
      - description: 账号
        in: query
        name: username
        required: true
        type: string
      - description: 密码
        in: query
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.UserLoginResponse'
      summary: 用户注册
      tags:
      - 用户
swagger: "2.0"